refactor: quebra ciclos e inverte camadas entre lib, components e ui
Prepara o grafo de dependências para a separação em pacotes, sem mover nada
para packages/ ainda. Só movimento de código e reescrita de import — nenhuma
mudança de lógica.

Ciclo api <-> http: ApiError e ApiErrorPayload passam a ser declarados em
lib/http/types.ts, a camada de baixo. lib/api/types.ts re-exporta os dois, então
nenhum dos ~15 consumidores foi tocado.

Inversão lib/ -> components/: o bloco de cálculo da lista de compra (ListItem e
as 15 funções que operam sobre ele) sai de components/domain/season/_shared.tsx
para lib/purchase-list/list-item.ts, e StageProductDraft sai de
timing-stages-editor.tsx para lib/timing/types.ts. Movi o módulo ListItem
inteiro, e não só o fecho mínimo, porque listItemToBuy e seedPlanOutputs chamam
listItemQuantity — cortar no fecho deixaria metade de um mesmo cálculo em cada
camada. _shared.tsx fica só com o shell de wizard em React.

Limpeza de components/ui: use-mobile.ts passa a viver dentro de components/ui, e
os 5 helpers puros de data (conversão local <-> YYYY-MM-DD, máscara e parse de
DD/MM/AAAA) saem de lib/timing/window-days.ts para lib/utils/dates.ts. Nenhum
deles carrega regra de janela de recomendação, então brazilian-date-input.tsx
deixa de depender de lib/timing. Em utils/dates.ts o format do date-fns entra
aliasado: formatDateBR já tem um parâmetro chamado format, que o sombrearia.

Depois disto: lib/ não importa de components/, components/ui não importa de
hooks/ nem de lib/timing, e http não depende de api.
