+++ src/lib/timing/window-days.ts
@@ -1,18 +1,10 @@
-import { addDays, differenceInCalendarDays, format } from "date-fns";
+import { addDays, differenceInCalendarDays } from "date-fns";
+import { dateToLocalYmd, localYmdToDate } from "@/lib/utils/dates";
 
 export const TIMING_WINDOW_TOLERANCE_DAYS = 2;
 
 /** Âncora de preview no editor de modelos (safra real recalcula na publicação). */
 export const TIMING_REFERENCE_YMD = "2026-01-01";
-
-export function localYmdToDate(ymd: string): Date {
-  const [year, month, day] = ymd.split("-").map((part) => Number.parseInt(part, 10));
-  return new Date(year, month - 1, day);
-}
-
-export function dateToLocalYmd(date: Date): string {
-  return format(date, "yyyy-MM-dd");
-}
 
 export function dayOffsetToIsoDate(dayOffset: number): string {
   const normalized = Number.isFinite(dayOffset) ? Math.round(dayOffset) : 0;
@@ -33,34 +25,6 @@ export function todayLocalYmd(): string {
   return dateToLocalYmd(new Date());
 }
 
-export function maskBrazilianDateInput(raw: string): string {
-  const digits = raw.replace(/\D/g, "").slice(0, 8);
-  if (digits.length <= 2) return digits;
-  if (digits.length <= 4) return `${digits.slice(0, 2)}/${digits.slice(2)}`;
-  return `${digits.slice(0, 2)}/${digits.slice(2, 4)}/${digits.slice(4)}`;
-}
-
-export function parseBrazilianDate(text: string): string | null {
-  const match = text.trim().match(/^(\d{2})\/(\d{2})\/(\d{4})$/);
-  if (!match) return null;
-
-  const day = Number.parseInt(match[1], 10);
-  const month = Number.parseInt(match[2], 10);
-  const year = Number.parseInt(match[3], 10);
-  if (month < 1 || month > 12 || day < 1 || day > 31 || year < 1900) return null;
-
-  const date = new Date(year, month - 1, day);
-  if (
-    date.getFullYear() !== year ||
-    date.getMonth() !== month - 1 ||
-    date.getDate() !== day
-  ) {
-    return null;
-  }
-
-  return dateToLocalYmd(date);
-}
-
 export function windowDatesFromRecommendedYmd(centerYmd: string): {
   startYmd: string;
   centerYmd: string;
@@ -75,10 +39,6 @@ export function windowDatesFromRecommendedYmd(centerYmd: string): {
     centerYmd: safeCenter,
     endYmd: dateToLocalYmd(addDays(center, TIMING_WINDOW_TOLERANCE_DAYS)),
   };
-}
-
-export function formatTimingPreviewDate(ymd: string): string {
-  return format(localYmdToDate(ymd), "dd/MM/yyyy");
 }
 
 export function targetDayToWindow(targetDay: number): {
